src/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.29.0
gunicorn==23.0.0
azure-identity>=1.17.0
aiohttp==3.11.1

# Semantic Kernel with proper packages
semantic-kernel>=1.0.0
semantic-kernel[azure_openai]
semantic-kernel[bing]
azure-ai-projects>=1.0.0b4
azure-ai-agents>=1.0.0b1

# Azure core dependencies  
azure-core==1.34.0
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry==1.6.9
opentelemetry-sdk
setuptools==80.9.0
starlette>=0.40.0 # fix vulnerability
jinja2 # new dependent of fastapi

# Additional dependencies
requests>=2.31.0